requirements.txt
accelerate==0.33.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
blis==0.7.11
catalogue==2.0.10
certifi==2024.7.4
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
confection==0.1.5
contourpy==1.2.1
cycler==0.12.1
cymem==2.0.8
datasets==2.21.0
deep-translator==1.11.4
deepspeed==0.14.5
dill==0.3.8
distro==1.9.0
emoji==2.12.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
faiss-cpu==1.8.0.post1
fastapi==0.115.0
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.1
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
fuzzywuzzy==0.18.0
greenlet==3.1.0
h11==0.12.0
h2==3.2.0
hjson==3.1.0
hpack==3.0.0
hstspreload==2024.8.1
httpcore==0.15.0
httpx==0.23.0
huggingface-hub==0.24.6
hyperframe==5.2.0
idna==2.10
itsdangerous==2.2.0
jieba==0.42.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
kiwisolver==1.4.5
langcodes==3.4.0
langdetect==1.0.9
language_data==1.2.0
Levenshtein==0.25.1
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
nvidia-ml-py==12.560.30
openai==1.42.0
packaging==24.1
pandas==2.2.2
peft==0.12.0
pgvector==0.3.3
pillow==10.4.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
preshed==3.0.9
psutil==6.0.0
psycopg2-binary==2.9.9
py-cpuinfo==9.0.0
pyarrow==17.0.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.25.1
pytz==2024.1
PyYAML==6.0.2
rapidfuzz==3.9.6
regex==2024.7.24
requests==2.32.3
rfc3986==1.5.0
rich==13.7.1
rouge==1.0.1
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.1
seaborn==0.13.2
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
soupsieve==2.6
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.35
srsly==2.4.8
starlette==0.38.6
sympy==1.13.2
thinc==8.2.5
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.0
tqdm==4.66.5
transformers==4.44.0
typer==0.12.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.31.0
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.0.4
wrapt==1.16.0
xxhash==3.5.0
yarl==1.9.4
